attrsview-app/src/androidTest/java/attrsview/app/SimpleViewTest.java
package attrsview.app;

import android.content.Context;
import android.support.test.InstrumentationRegistry;
import android.support.test.rule.ActivityTestRule;
import android.support.test.runner.AndroidJUnit4;
import attrsview.app.views.SimpleView;
import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;

import static org.junit.Assert.assertEquals;

/**
 * Instrumentation test, which will execute on an Android device.
 *
 * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
 */
@RunWith(AndroidJUnit4.class) public class SimpleViewTest {

  @Rule public ActivityTestRule<MainActivity> mActivityRule =
      new ActivityTestRule<>(MainActivity.class);

  @Test public void useAppContext() throws Exception {
    // Context of the app under test.
    Context appContext = InstrumentationRegistry.getTargetContext();
    assertEquals("attrsview.app", appContext.getPackageName());
  }

  @Test public void getTextTest() {
    SimpleView simpleView = (SimpleView) mActivityRule.getActivity().findViewById(R.id.simple_view);
    assertEquals(simpleView.getText(), "Attrs View");
  }
}
